+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Backwards incompatible
 
 * Removed `conman.pages` app.
-* Removed `slug` and `parent` fields.
+* Removed `slug` and `parent` fields from Route.
 * Removed dependency upon `django-polymorphic-tree`.
 * Destroyed and recreated migrations.
 * Removed `conman.routes.handlers.SimpleHandler`.
@@ -27,6 +27,15 @@
   represent a URL that redirects to another URL (that is not represented by a
   `Route`). One might expect this to be a URL on an external site.
 * Allow `Route` to delegate model checks to the associated handler.
+
+### Added
+
+* `Route.objects.move_branch()` can be used to move a `Route` and its
+  descendants to another location.
+* `Route().move_to(new_url, *, move_children)` can be used to move a `Route`
+  and (optionally) it's descendants to another location.
+* `Route().swap_with(new_url, *, move_children)` can be used to swap a `Route`
+  and (optionally) it's descendants with another `Route`.
 
 ### Changed